Change async boot up to class async boot up
